+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Rename the test files
 - New test: version command
 - Implement the version command
+- New test: replace command
 
 ## 0.0.1 - 2022-02-01
 - Start protocol information details